Parameter & UI Update!

+++ delete.php
@@ -4,29 +4,34 @@ session_start();
 
 // Redirect to index.html if not logged in
 if (!isset($_SESSION['user_id'])) {
-    header("Location: index.html"); // Redirect to home page
+    header("Location: login.php"); // Redirect to home page
     exit();
 }
 
+
 $usr_id = $_SESSION['user_id'];
-$baseDirs = ["/var/www/html/MyNAS/uploads/$usr_id", "/var/www/html/MyNAS/backup/$usr_id"];
+$baseDirs = ["/var/www/MyNAS/uploads/$usr_id", "/var/www/MyNAS/backup/$usr_id"];
 
 
 if ($_SERVER["REQUEST_METHOD"] == "POST" && isset($_POST["file"]) && isset($_POST["folder"])) {
-    $file = basename($_POST["file"]);
+    $file = stripslashes($_POST["file"]);
+    $file = urldecode($file);
+    $file = basename($file);
+    $file = str_replace("'", "\'", $file);
     $folder = rtrim($_POST["folder"], "/") . "/";
-    $filePath = realpath($folder . $file);
+    $safeFile = escapeshellarg($file);
+    $filePath = realpath($folder . $safeFile);
 
     $isValidPath = false;
 
 	foreach ($baseDirs as $baseDir) {
-        	if($filePath !== false && strpos($filePath, $baseDir) === 0) {
+        	if($filePath !== false && str_starts_with($filePath, $baseDir)) {
                 	$isValidPath = true;
 			break;
         	}
 	}
 
-    if ($isValidPath && file_exists($filePath)) {
+    if ($isValidPath && file_exists($filePath) && is_file($filePath)) {
         if (unlink($filePath)) {
             header("Location: MyFiles.php?msg=File deleted successfully");
             exit();
